Add vim-plug and solarized to setup
Vim plug seems to be a slightly more intelligent way to handle plugins,
if the lazy loading mechanism is truly as effective as claimed.

The syntax for it seems nice and I don't have to worry about Ruby
scripts anymore, so that's enough for me at the moment.

Additionally, giving solarized a try for my terminal and Vim, and
liking it so far.